Simplify %e logic; fix trailing 0s issue.
Previously, provided two patterns, for when the absolute value was
greater/less than one.  Now just provide one pattern but branch for the
e+ / e- choice.

Additionally, correctly ensure that we get the right number of requested
digits in the case when there are trailing zeros that are otherwise
truncated.